fix assertions
